+++ CHANGELOG.md
@@ -2,10 +2,17 @@
 
 All Notable changes to `Laravel AdWords Targeting Idea Service` will be documented in this file.
 
-## 1.1.0
-- Added Configurable SoapLogger
+## [1.2.0] - 2017-07-25
+### Added
+- Fluent method to convert null values to 0 or empty collection
+### Changed
+- Using version 28.0.0 of googleads/googleads-php-lib
+
+## [1.1.0] - 2017-06-20
+### Added
+- Configurable SoapLogger
 - Console Command For Generating Refresh Token
 - Auto Discovery For L5.5
 
-## 1.0.0
+## [1.0.0] - 2017-06-06
 - Initial Release